[#4] Fix : Diary - long형 Long으로 수정 - create(), update() 메소드 수정

+++ src/main/java/mashup/backend/etc/our/diary/repository/DiaryRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface DiaryRepository extends JpaRepository<Diary, Integer> {
+public interface DiaryRepository extends JpaRepository<Diary, Long> {
     Optional<Diary> findByDiaryId(Long diaryId);
 //    List<Diary> findByUserIdAndGroupId(Long userId, Long groupId);
 }
